requirements.txt
Flask
Flask-SQLAlchemy
Flask-Caching
pymysql
geoip2
pip
python-dotenv